+++ src/cursor.h
@@ -5,7 +5,7 @@
 class Cursor
 {
 public:
-	Cursor(int width, int height, int max_row, int max_col);
+	Cursor(int width, int height, size_t max_row, size_t max_col);
 
 	void moveDown(size_t buffer_max_col);
 
@@ -15,23 +15,23 @@ public:
 
 	void moveRight(size_t line_max_row);
 
-	int row();
+	size_t row();
 
-	int col();
+	size_t col();
 
 	int x();
 
 	int y();
 
-	void move(int row, int col);
+	void move(size_t row, size_t col);
 
 	void draw();
 
 private:
-	int m_max_col;
-	int m_max_row;
-	int m_row;
-	int m_col;
+	size_t m_max_col;
+	size_t m_max_row;
+	size_t m_row;
+	size_t m_col;
 	int m_x;
 	int m_y;
 	int m_width;
